# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Olena Studio",
    "version": "1.0.9",
    "description": "home project",
    "contact": {
      "name": "Maria Skydan",
      "email": "mariaskydan@gmail.com",
      "url": "https://github.com/maskaa22?page=1&tab=repositories"
    }
  },
  "parameters": {
    "page": {
      "name": "page",
      "in": "query",
      "type": "number",
      "description": "Введіть сторінку продуктів",
      "default": 2
    },
    "limit": {
      "name": "limit",
      "in": "query",
      "type": "number",
      "description": "Введіть максимальну кількість продуктів на сторінці",
      "default": 5
    },
...
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "product"
        ],
        "operationId": "getAllProducts",
        "summary": "Перегляд списку продуктів",
        "parameters": [
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/categorySchemaWithPage"
            }
          }
        }
      },
      "post": {
        "tags": [
          "product"
        ],
        "operationId": "createProduct",
        "summary": "Створення продукту",
        "parameters": [
          {
            "name": "product_name",
            "in": "formData",
            "type": "string",
            "description": "Назва товару",
            "default": "Balsam peptide",
            "required": true
          },
          {
            "name": "title",
            "in": "formData",
            "type": "string",
            "description": "Короткий опис товару",
            "default": "Бальзам-пополіскувач з пептидами",
            "required": true
          },
          {
            "name": "dosage",
            "in": "formData",
            "type": "string",
            "description": "Дозування товару в мл",
            "default": "200",
            "required": true
          },
          {
            "name": "price",
            "in": "formData",
            "type": "string",
            "description": "Ціна товару за 1 шт.",
            "default": "150",
            "required": true
          },
          {
            "name": "inventoryNumber",
            "in": "formData",
            "type": "string",
            "description": "Інвентарний номер товару",
            "default": "39",
            "required": true
          },
          {
            "name": "category_id",
            "in": "formData",
            "type": "string",
            "description": "Ідентифікатор категорії в яку відноситься товар",
            "default": "63fa1ba717474ba5104f9fc5",
            "required": true
          },
          {
            "name": "img",
            "in": "formData",
            "type": "file",
            "description": "Завантажити фото товару",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/productSchema",
              "type": "object"
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "$ref": "#/definitions/errorSchemaInventoryNumberCheck"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaCheckData"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "product"
        ],
        "operationId": "deleteProduct",
        "summary": "Видалення продукту",
        "parameters": [
          {
            "name": "inventoryNumber",
            "in": "query",
            "type": "string",
            "description": "Інвентарний номер товару",
            "default": "39",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/deleteSchema"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaInventoryNumberInvalid"
            }
          }
        }
      }
    },
...
    "/products/:user_id/orders": {
      "get": {
        "tags": [
          "product"
        ],
        "operationId": "getOrderByIdUser",
        "summary": "Список замовлень користувача по його id",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Ідентифікатор користувача",
            "default": "64159028f74553185be6c167"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ordersSchema"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundData"
            }
          }
        }
      }
    },
    "/products/orders_filter": {
      "get": {
        "tags": [
          "product"
        ],
        "operationId": "getOrderByIdUser",
        "summary": "Фільтрування списку замовлень статусу його обробки",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Статус",
            "default": "готово"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ordersSchema"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundData"
            }
          }
        }
      }
    },
...
    "/products/orders": {
      "get": {
        "tags": [
          "product"
        ],
        "operationId": "getAllOrders",
        "summary": "Список всіх замовлень",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ordersSchema"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundOrders"
            }
          }
        }
      },
      "post": {
        "tags": [
          "product"
        ],
        "operationId": "createOrder",
        "summary": "Створення замовлення користувачем",
        "description": "Вводяться дані користувача та продуктів і статус(очікується, прийнято, відправлено, готово) з малої літери",
        "parameters": [
          {
            "name": "createOrder",
            "in": "body",
            "schema": {
              "type": "object",
              "default": {
                "user_id": "64159105f74553185be6c182",
                "user_name": "Луі",
                "surname": "Сальваторе",
                "phone": "0987793157",
                "nameSity": "Вінниця",
                "nameDepartment": "25",
                "pay": "Готівка",
                "cart": [
                  {
                    "_id": "555df2q5dcf",
                    "product_name": "Orising Lozione Tonica Caduta",
                    "title": "Ампули проти випадіння",
                    "dosage": "10х12",
                    "price": "256",
                    "img": "67e0b763-6316-4cc3-b7f4-6298f435f747.jpg",
                    "category_id": "63fa1be517474ba5104f9fce",
                    "count": "1",
                    "totalPrice": "256",
                    "inventoryNumber": "25",
                    "createdAt": "2023-12-13T10:23:01.414Z",
                    "updatedAt": "2023-12-13T10:24:38.892Z",
                    "__v": 0,
                    "id": "555df2q5dcf"
                  }
                ],
                "status": "очікується",
                "summa": "402",
                "month": "грудень"
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ordersSchema"
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFound"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundUser"
            }
          },
          "409": {
            "description": "Error: Conflict",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "example": "Помилка оформлення замовлення"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "product"
        ],
        "operationId": "updateOrderStatus",
        "summary": "Зміна статусу оформлення замовлення",
        "parameters": [
          {
            "name": "updateStatus",
            "in": "body",
            "description": "Введіть ідентифікатор замовлення і статус(очікується, прийнято, відправлено, готово) з малої літери",
            "schema": {
              "type": "object",
              "default": {
                "_id": "6432a0c5f46a316e0ead7541",
                "status": "прийнято"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/updateSchema"
            }
          },
          "409": {
            "description": "Error: Conflict",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "example": "Статус не змінений"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "errorSchemaCheckData": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Перевірте дані"
        }
      }
    },
    "errorSchemaNotFound": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Заповніть поля"
        }
      }
    },
    "errorSchemaNotFoundData": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Дані відсутні"
        }
      }
    },
    "errorSchemaNotFoundOrders": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Не має замовлень"
        }
      }
    },
    "errorSchemaNotFoundUser": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Користувач відсутній"
        }
      }
    },
...
    "errorSchemaInventoryNumberCheck": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Інвентарний номер вже використаний"
        }
      }
    },
    "errorSchemaInventoryNumberInvalid": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Не правильний інвентарний номер"
        }
      }
    },
...
    "ordersSchema": {
      "type": "array",
      "items": {
        "properties": {
          "_id": {
            "description": "Mongo ідентифікатор",
            "example": "555df2q5dcf"
          },
          "user_id": {
            "description": "Mongo ідентифікатор",
            "example": "555df2q5dcfsdfe56ygdfb"
          },
          "user_name": {
            "description": "Ім'я користувача",
            "type": "string",
            "example": "Луі"
          },
          "surname": {
            "description": "Прізвище користувача",
            "type": "string",
            "example": "Сальваторе"
          },
          "phone": {
            "description": "Номер телефону",
            "type": "string",
            "example": "0987793157"
          },
          "nameSity": {
            "description": "Місто одержання посилки для Нової пошти",
            "type": "string",
            "example": "Вінниця"
          },
          "nameDepartment": {
            "description": "Номер відділення Нової пошти",
            "type": "string",
            "example": "25"
          },
          "pay": {
            "description": "Вид сплати (готівка/карта)",
            "type": "string",
            "example": "Готівка"
          },
          "cart": {
            "items": {
              "$ref": "#/definitions/productSchema"
            }
          },
          "status": {
            "description": "Статус обробки замовлення",
            "type": "string",
            "example": "готово"
          },
          "summa": {
            "description": "Сума до сплати замовником",
            "type": "string",
            "example": "402"
          },
          "month": {
            "description": "Назва місяця даного замовлення",
            "type": "string",
            "example": "грудень"
          },
          "createdAt": {
            "description": "Дата створення користувача",
            "type": "string",
            "example": "2023-12-13T10:23:01.414Z"
          },
          "updatedAt": {
            "description": "Дата внесення змін",
            "type": "string",
            "example": "2023-12-13T10:24:38.892Z"
          },
          "__v": {
            "description": "Версійність в Mongo",
            "type": "integer",
            "example": 0
          },
          "id": {
            "description": "Mongo ідентифікатор",
            "example": "555df2q5dcf"
          }
        }
      }
    },
...
    "categorySchemaWithPage": {
      "type": "object",
      "properties": {
        "docs": {
          "description": "Список товарів",
          "$ref": "#/definitions/productsSchema"
        },
        "total": {
          "type": "number",
          "description": "Загальна кількість товарів по заданій категорії",
          "example": 12
        },
        "limit": {
          "type": "number",
          "description": "Максимальна кількість товарів на сторінці",
          "example": "555df2q5dcf"
        },
        "page": {
          "type": "number",
          "description": "Номер сторінки",
          "example": "555df2q5dcf"
        },
        "pages": {
          "type": "number",
          "description": "Кількість всіх сторінок",
          "example": "555df2q5dcf"
        }
      }
    },
    "deleteSchema": {
      "type": "object",
      "properties": {
        "deletedCount": {
          "type": "number",
          "description": "Текст повідомлення",
          "example": 1
        }
      }
    },
    "updateSchema": {
      "type": "object",
      "properties": {
        "acknowledged": {
          "type": "boolean",
          "example": true
        },
        "modifiedCount": {
          "type": "number",
          "example": 1
        },
        "upsertedId": {
          "example": null
        },
        "upsertedCount": {
          "type": "number",
          "example": 0
        },
        "matchedCount": {
          "type": "number",
          "example": 1
        }
      }
    },
...
    "productSchema": {
      "properties": {
        "_id": {
          "description": "Mongo ідентифікатор",
          "example": "555df2q5dcf"
        },
        "product_name": {
          "description": "Назва товару",
          "type": "string",
          "example": "Orising Lozione Tonica Caduta"
        },
        "title": {
          "description": "Короткий опис товару",
          "type": "string",
          "example": "Ампули проти випадіння"
        },
        "dosage": {
          "description": "Дозування продукту",
          "type": "string",
          "example": "10х12"
        },
        "price": {
          "description": "Ціна 1-ї шт. даного товару",
          "type": "string",
          "example": "256"
        },
        "img": {
          "description": "Назва файлу фотокартки",
          "type": "string",
          "example": "67e0b763-6316-4cc3-b7f4-6298f435f747.jpg"
        },
        "category_id": {
          "description": "Ідентифікатор категорії",
          "type": "string",
          "example": "63fa1be517474ba5104f9fce"
        },
        "count": {
          "description": "Кількість шт. даного товару",
          "type": "string",
          "example": "1"
        },
        "totalPrice": {
          "description": "Загальна сума за товар",
          "type": "string",
          "example": "256"
        },
        "inventoryNumber": {
          "description": "Номер відділення Нової пошти",
          "type": "string",
          "example": "25"
        },
        "createdAt": {
          "description": "Дата створення продукту",
          "type": "string",
          "example": "2023-12-13T10:23:01.414Z"
        },
        "updatedAt": {
          "description": "Дата внесення змін",
          "type": "string",
          "example": "2023-12-13T10:24:38.892Z"
        },
        "__v": {
          "description": "Версійність в Mongo",
          "type": "integer",
          "example": 0
        },
        "id": {
          "description": "Mongo ідентифікатор",
          "example": "555df2q5dcf"
        }
      }
    },
    "productsSchema": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/productSchema"
      }
    },
...
  }
}